Added getTokens util and moved common steps
