+++ features/catalog/step_definitions/e-service-consumers-download.ts
@@ -3,10 +3,11 @@ import { z } from "zod";
 import {
   assertContextSchema,
   getAuthorizationHeader,
+  getToken,
 } from "../../../utils/commons";
 import { dataPreparationService } from "../../../services/data-preparation.service";
 import { apiClient } from "../../../api";
-import { Party } from "./common-steps";
+import { Party } from "../../common-steps";
 
 Given(
   "{string} ha un agreement attivo con quell'e-service",
@@ -15,7 +16,7 @@ Given(
       eserviceId: z.string(),
       descriptorId: z.string(),
     });
-    const token = this.tokens[party]!.admin!;
+    const token = getToken(this.tokens, party, "admin");
     const agreementId = await dataPreparationService.createAgreement(
       token,
       this.eserviceId,
